Create manatee crate, convert main project into manatee_editor
