+++ crates/manatee/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "manatee"
+description = "A ridiculously simple Rust-based game engine"
+version = "0.0.1"
+edition = "2021"
+
+[dependencies]
+winit = "0.30.0"
