+++ src/main.rs
@@ -1,42 +1,5 @@
-use winit::{
-    application::ApplicationHandler,
-    event::WindowEvent,
-    event_loop::{ActiveEventLoop, ControlFlow, EventLoop},
-    window::{Window, WindowId},
-};
-
-#[derive(Default)]
-struct App {
-    window: Option<Window>,
-}
-
-impl ApplicationHandler for App {
-    fn window_event(&mut self, event_loop: &ActiveEventLoop, _id: WindowId, event: WindowEvent) {
-        // This will probably need to be a match after the app grows a bit
-        if event == WindowEvent::CloseRequested {
-            event_loop.exit();
-        }
-    }
-    fn resumed(&mut self, event_loop: &ActiveEventLoop) {
-        self.window = Some(
-            event_loop
-                .create_window(Window::default_attributes())
-                .unwrap(),
-        );
-    }
-}
-
-// TODO: This and the App struct should probably be abstracted to their own crate at some point
-pub fn create_and_run_window() {
-    let event_loop = EventLoop::new().unwrap();
-
-    // ControlFlow::Poll continuously runs the event loop, even if the OS hasn't
-    // dispatched any events. This is ideal for games and similar applications.
-    event_loop.set_control_flow(ControlFlow::Poll);
-
-    let mut app = App::default();
-    let _ = event_loop.run_app(&mut app);
-}
+#![windows_subsystem = "windows"]
+pub use manatee::window::create_and_run_window;
 
 pub fn main() {
     create_and_run_window();
